% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{contactinfo}
\alias{contactinfo}
\title{Wrapper around \code{apputils::contactinfo}}
\usage{
contactinfo(id = "leonawicz", snap, iarc, uaf)
}
\arguments{
\item{id}{character, name of author template to use. Available templates: \code{"leonawicz"}.}

\item{snap}{image url.}

\item{iarc}{image url.}

\item{uaf}{image url.}
}
\value{
a tag list.
}
\description{
This function makes a call to \code{apputils::contactinfo} using stored author contact info templates.
This override requires that \code{snaputils} be loaded after \code{apputils}, but this is the meaningful
order for loading these packages by their nature.
Otherwise, \code{snapputils::contactinfo} would have to be called explicitly.
}
\details{
The user may optionally provide paths to SNAP, IARC and UAF logo images for branding.
The logos may be contained in the \code{www/} directory or could be at a web url.
}
\examples{
\dontrun{contactinfo()}
}
