% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faq.R
\name{faq}
\alias{faq}
\title{Generate a SNAP Frequently Asked Questions (FAQ) widget}
\usage{
faq(id, format = "bscollapse", bscollapse_args = list(id = "faq", open =
  NULL, multiple = FALSE), showcase_args = list(drop = NULL))
}
\arguments{
\item{id}{character, the id for the \code{shinyBS::bsCollapse} element.}

\item{format}{character, \code{"bscollapse"} (default) or alternatively, \code{"list"}.}

\item{bscollapse_args}{a list of arguments passed to bsCollapse. Defaults to \code{list(id = "faq", open = NULL, multiple = FALSE)}.}

\item{showcase_args}{a list of arguments passed to \code{snapp_showcase}.}
}
\value{
a \code{shinyBS::bsCollapse} object or a list.
}
\description{
Generate a widget that displays FAQs in an information/about section of an SNAP Shiny app.
}
\details{
\code{faq} is a simple function containing a dictionary of common SNAP FAQ entries.
It is meant to override \code{apputils::faq}.
It returns a Bootstrap Collapse menu FAQ format by default. Changing to \code{format = "list"}
alternatively returns a list of question and answer pairs, each being a length-2 sublist.
In this case, each question is a character string and each answer is either a simple HTML paragraph
or a taglist of multiple HTML components if the answer contains more complex content (e.g., images and links).

A key difference with the \code{open} argument that can be passed in a list to \code{bscollapse_args}
is that it must simply match a single \emph{dictionary entry} \code{id}, not a collapse panel question title or id (e.g., \code{open = "apps"}).

The current FAQ dictionary IDs and the questions they refer to include:
\describe{
  \item{distributions}{Why use probability distributions?}
  \item{gcm}{What is a GCM?}
  \item{rcp}{What is an RCP?}
  \item{fmz}{What are fire management zones?}
  \item{apps}{How did you make this app? Are other apps available?}
  \item{alfresco_simulations}{What do "simulations" refer to in the data selection area?}
  \item{factsheet_about}{What are fact sheets?}
  \item{factsheet_availability}{Why are fact sheets not always available?}
  \item{climdist_variables}{What do the available distributions represent?}
}
}
\examples{
#not run
}
