% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snaputils.R
\name{snap_res}
\alias{snap_res}
\title{Get a snaputils resource path}
\usage{
snap_res(type = "images")
}
\arguments{
\item{type}{character, resource type. \code{"images"} is the only currently available resource type.}
}
\description{
Get a resource path in an app to local package resources.
}
\details{
This function is called in the UI code of an app that relies \code{snaputils} package resources, typically image logos.
}
\examples{
\dontrun{snap_res()}
}
