% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{snapp_showcase}
\alias{snapp_showcase}
\title{Genrate SNAP apps showcase content}
\usage{
snapp_showcase(drop = NULL)
}
\arguments{
\item{drop}{character vector of any apps to exclude from showcase.}
}
\value{
a shiny fluidRow containing organized and stylized app image links for reference.
}
\description{
This is a wrapper around \code{apputils::app_showcase}.
}
\examples{
# not run
}
