% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{snapp_titles}
\alias{snapp_titles}
\title{Get heading information related to SNAP Shiny apps}
\usage{
snapp_titles(type)
}
\arguments{
\item{type}{character, \code{"label"}, \code{"title"} or \code{"sub"}.}
}
\value{
a vector of headings.
}
\description{
This function returns a list of headings related to SNAP Shiny apps.
}
\details{
Labels are generally shown above an app showcase widget. Titles are generally the top line heading over the image link inside the widget.
Subtitles are for smaller text below.
}
\examples{
snapp_titles("label")
}
