% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{snapp_urls}
\alias{snapp_urls}
\title{Get urls related to SNAP Shiny apps}
\usage{
snapp_urls(type = "app")
}
\arguments{
\item{type}{character, \code{"app"}, \code{"img"} or \code{"gh"}.}
}
\value{
a vector of url strings.
}
\description{
This function returns a list of urls related to SNAP Shiny apps.
}
\details{
App urls are for the hosted apps. Image urls are to thumbnail images representing each app.
GitHub urls are to the code repositories related to each app.
}
\examples{
snapp_urls()
}
