% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snaputils.R
\docType{package}
\name{snaputils}
\alias{snaputils}
\alias{snaputils-package}
\title{snaputils: SNAP-specific utilities for Shiny apps.}
\description{
The \code{snaputils} package contains utilities for Shiny app development.
The package inherits primarily from `apputils`, which contains common utility functions, settings and references for use across multiple Shiny apps.
It also imports \code{maputils}. As the name suggests, \code{snaputils} is specific to a SNAP context. So is \code{maaputils} though separate.
Only \code{apputils} has a more general non-SNAP context.
All three packages are satellite members of the SNAPverse collection of R packages.
}
