% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snaputils.R
\docType{package}
\name{snaputils}
\alias{snaputils}
\alias{snaputils-package}
\title{SNAP-specific utilities for Shiny apps.}
\description{
The \code{snaputils} package common utility functions, settings and references used across multiple Shiny apps
specific to a SNAP context. This package inherents from the apputils package and is a satellite member of the SNAPverse collection of R packages.
}
